+++ src/Resources/app/administration/src/module/swag-paypal/components/swag-paypal-checkout/index.js
@@ -263,6 +263,7 @@ Component.register('swag-paypal-checkout', {
                     return response.filter((paymentMethod) => {
                         return paymentMethod.formattedHandlerIdentifier !== 'handler_swag_pospayment'
                             && paymentMethod.formattedHandlerIdentifier !== 'handler_swag_trustlyapmhandler'
+                            && paymentMethod.formattedHandlerIdentifier !== 'handler_swag_giropayapmhandler'
                             && paymentMethod.formattedHandlerIdentifier !== 'handler_swag_sofortapmhandler';
                     });
                 });