+++ src/Storefront/Data/Service/SPBCheckoutDataService.php
@@ -25,9 +25,7 @@ class SPBCheckoutDataService extends AbstractCheckoutDataService
 {
     private const APM_BLIK = 'blik';
     private const APM_EPS = 'eps';
-    private const APM_GIROPAY = 'giropay';
     private const APM_P24 = 'p24';
-    private const APM_SOFORT = 'sofort';
 
     private SystemConfigService $systemConfigService;
 
@@ -99,8 +97,6 @@ class SPBCheckoutDataService extends AbstractCheckoutDataService
 
         if ($totalPrice < 1.0 && $currencyIsoCode === 'EUR') {
             $disabled[] = self::APM_EPS;
-            $disabled[] = self::APM_GIROPAY;
-            $disabled[] = self::APM_SOFORT;
         }
 
         if ($totalPrice < 1.0 && $currencyIsoCode === 'PLN') {