+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - PPI-896 - Changed the wording of the checkout Smart Payment Buttons from 'PayPal Checkout' to 'Pay with PayPal'.
 - PPI-932 - Removed the no longer supported Sofort payment method
 - PPI-936 - Fixes an issue, where Zettle may receive incorrect tax rates
+- PPI-948 - Removed the no longer supported Giropay payment method
 
 # 6.4.0
 - PPI-853 - Added option to add buyer country to "Pay Later" banners